relay-dispatch: split relay-manifest into narrower platform boundaries (#188)
Split 1836-line `skills/relay-dispatch/scripts/relay-manifest.js` into seven
slice modules under `skills/relay-dispatch/scripts/manifest/`:

- manifest/paths.js        — run-id, slug, trust-root, validateManifestPaths
- manifest/store.js        — frontmatter codec + read/write/list/skeleton
- manifest/lifecycle.js    — STATES, ALLOWED_TRANSITIONS, validateTransition,
                             forceTransitionState, validateTransitionInvariants
- manifest/rubric.js       — getRubricAnchorStatus, grandfather gate,
                             symlink-safe I/O, migration-manifest cross-check
- manifest/cleanup.js      — cleanup skeleton, runCleanup, git helpers
- manifest/attempts.js     — readPreviousAttempts, captureAttempt
- manifest/environment.js  — collectEnvironmentSnapshot, compareEnvironmentSnapshot

relay-manifest.js becomes a pure re-export facade (17 lines, zero function
declarations) so test files, test-support.js, relay-request.js, and plan-side
helpers keep working without edits.

Nine runtime entry points migrate to slice-specific imports:
dispatch.js, recover-state.js, close-run.js, update-manifest-state.js,
relay-resolver.js, review-runner.js, review-gate.js, gate-check.js,
finalize-run.js. Additional runtime consumers (relay-events.js,
relay-migrate-rubric.js, reliability-report.js, cleanup-worktrees.js)
also migrated to narrower imports.

Gate-layer behavior byte-identical: validateTransition matrix, forceTransitionState
caller set (recover-state.js + relay-migrate-rubric.js), getRubricAnchorStatus
on forged/missing/symlinked manifests. External verifier reference
(~/.relay/migrations/rubric-mandatory.yaml via loadMigrationManifest) preserved
as single definition.

Tests: 444 → 458 (+14). New manifest/*.test.js + manifest-direct-imports.test.js
exercise the narrower modules with direct imports (no facade).

Trust-model audit:
- Q1 (forge): gate-layer-behavior-byte-identical — no duplicate gate functions.
- Q2 (gate): five sites named in docs/issue-188-manifest-boundary-split.md.
- Q3 (external verifier): rubric-mandatory.yaml unchanged code path.

Deferred: #189 (review-runner), #190 (grandfather retirement), #191 (resolver
hygiene). relay-intake/scripts/relay-request.js migration deferred per plan.

Closes #188
