+++ skills/relay-dispatch/scripts/manifest-direct-imports.test.js
@@ -1,0 +1,7 @@
+require("./manifest/paths.test");
+require("./manifest/store.test");
+require("./manifest/lifecycle.test");
+require("./manifest/rubric.test");
+require("./manifest/cleanup.test");
+require("./manifest/attempts.test");
+require("./manifest/environment.test");
